+++ src/react-app-env.d.ts
@@ -80,3 +80,4 @@ declare module '*.module.sass' {
   const classes: { readonly [key: string]: string }
   export default classes
 }
+declare module 'crypto-js'